+++ lib/core/constants/strings.dart
@@ -1,5 +1,5 @@
 class AppStrings {
   // App
   static const appTitle = 'Fintek';
-  static const appVersion = 'Version: 1.0.1';
+  static const appVersion = 'Version: 1.0.4';
 }